requirements.txt
flask
twilio
python-dotenv
selenium
webdriver-manager
pandas
requests
beautifulsoup4